Align enums to recent object changes

+++ adl/test/scenarios/v2/multiple/apimanagement/output/anonymous/alias_414.ts
@@ -1,1 +1,1 @@
-export type alias_414 = double;
+export type alias_414 = int32;
